added permission request

+++ src/main/java/com/example/backend_academic_monitoring/Implementations/ImageServiceImpl.java
@@ -7,48 +7,51 @@ import com.example.backend_academic_monitoring.Repository.ImageRepository;
 import com.example.backend_academic_monitoring.Service.ImageService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.ByteArrayResource;
-import org.springframework.core.io.InputStreamResource;
 import org.springframework.core.io.Resource;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.util.UUID;
 
 @Service
 public class ImageServiceImpl implements ImageService {
-    private final ImageRepository imageRepository;
     public static final Logger LOGGER = LoggerFactory.getLogger(ImageServiceImpl.class);
+    private final ImageRepository imageRepository;
+    @Value("${server.host}")
+    private String host;
+    @Value("${server.port}")
+    private String port;
 
     public ImageServiceImpl(ImageRepository imageRepository) {
         this.imageRepository = imageRepository;
     }
 
     @Override
-    public Integer saveFile(MultipartFile file) {
+    public ImageEntity saveFile(MultipartFile file) {
         UUID uuid = UUID.randomUUID();
         String userDirectoryPath = System.getProperty("user.dir");
-        LOGGER.info("userDirectoryPath {}",userDirectoryPath);
-        if(!new File(userDirectoryPath + "/images").exists()){
-            new File(userDirectoryPath + "/images").mkdir();
+        LOGGER.info("userDirectoryPath {}", userDirectoryPath);
+        if (!new File(userDirectoryPath + "/images").exists()) {
+            LOGGER.info("Creating directory");
+            if (new File(userDirectoryPath + "/images").mkdir())
+                LOGGER.info("Directory created");
+            else
+                throw new RuntimeException("Error trying to create directory");
         }
-
         try {
-            ImageEntity imageDTO = new ImageEntity();
-            imageDTO.setUuid(uuid.toString());
-            imageDTO.setType(file.getContentType());
-            imageDTO.setName(file.getOriginalFilename());
-            LOGGER.info("DTO {}",imageDTO);
-            imageDTO = imageRepository.save(imageDTO);
-            file.transferTo(new File(userDirectoryPath +"/images/" + imageDTO.getUuid()));
-            return imageDTO.getId();
+            ImageEntity imageEntity = new ImageEntity();
+            imageEntity.setUuid(uuid.toString());
+            imageEntity.setType(file.getContentType());
+            imageEntity.setName(file.getOriginalFilename());
+            LOGGER.info("DTO {}", imageEntity);
+            imageEntity = imageRepository.save(imageEntity);
+            file.transferTo(new File(userDirectoryPath + "/images/" + imageEntity.getUuid()));
+            return imageEntity;
         } catch (IOException ex) {
             throw new RuntimeException("Archivo Invalido");
         }
@@ -80,12 +83,15 @@ public class ImageServiceImpl implements ImageService {
     @Override
     public void deleteImage(Integer userId) {
         ImageEntity imageEntity = imageRepository.findByUserId(userId);
-        LOGGER.info("imageEntity {}",imageEntity);
+        LOGGER.info("imageEntity {}", imageEntity);
         String userDirectoryPath = System.getProperty("user.dir");
         File file = new File(userDirectoryPath + "/images/" + imageEntity.getUuid());
-        file.delete();
-        imageRepository.delete(imageEntity);
+        if (file.delete())
+            imageRepository.delete(imageEntity);
+        else
+            throw new RuntimeException("Error trying to delete file");
     }
+
     public ResourceDTO getResource(String uuid) throws IOException {
         ResourceDTO resourceDTO = new ResourceDTO();
         ImageDTO imageDTO = getImageByUuid(uuid);
@@ -105,5 +111,10 @@ public class ImageServiceImpl implements ImageService {
         imageDTO.setType(imageEntity.getType());
         imageDTO.setUuid(imageEntity.getUuid());
         return imageDTO;
+    }
+
+    @Override
+    public String getImageURL(String uuid) {
+        return "http://" + host + ":" + port + "/file/image/" + uuid;
     }
 }
